// src/shared/models/driver.ts
export type Driver = {
  id?: number;
  name: string;
  surname: string;
  championship_name: string;
  category: string;
  team_name: string;
  end_year: number;
  expected_standing: number;
  team_loyalty: number;
  country: string;
  dob: string;
  elo: number;
  race_skill: number;
  qualifying_skill: number;
  aggression: number;
  defending: number;
  stamina: number;
  consistency: number;
  start_reactions: number;
  wet_skill: number;
  tyre_management: number;
  fuel_management: number;
  blue_flag_conceding: number;
  weather_tyre_changes: number;
  avoidance_of_mistakes: number;
  avoidance_of_forced_mistakes: number;
  setup_downforce: number;
  setup_downforce_randomness: number;
};